Moves state and methods from App component to the DeckBuilder component
The selectedCards state inside the App component was only used by the
Deckbuilder component. Therefore, I decided to move that state and any
associated methods to the Deckbuilder component as it is the top-most
relevant component.

The editDeck method inside the App component, however, is now broken as
it would set the state of the selectedCards whenever a user selected a
specific deck. Now I will work on a solution for that problem that will
accomodate this change. Or, if the change seems irrational later on, I
will revert it.
